Update forum-moderation.adoc
Seems like there's too much overlap between the Home and Moderate tabs, and the current doc is too redundant as a result. I'd also like to know more about what's on the Rules tab, and details about what shows up on the sub-screens when a user clicks something like "Frozen Members". We could use that info to construct more of a walkthrough and less of a description.